seperated arp and dns hunters, made them active and fixed some coe on arp
